Added number tests at region level for Belgium
Belgium has 3 regions : Flanders, Wallonia and Brussels. Flanders is in the north of Belgium we people talk Dutch, Wallonia in the south were people talk French and Brussels in the center is bilangual. Next we have 11 provinces . So we have the 'all' key for Belgium, 3 keys for the regions and then 10 keys for the 11 provinces (Brussels is already has a region key). In total 14 keys. More info here: https://en.wikipedia.org/wiki/Provinces_of_Belgium#List